Add basic spot checker
