+++ lib/constants.js
@@ -1,2 +1,3 @@
+export const MAX_CAR_SIZE = 200 * 1024 * 1024 // 200 MB
 export const RPC_URL = 'https://api.node.glif.io/'
 export const RPC_AUTH = 'cL/ik5kups4LCAg6a+s6CS40/JcARERwZ4uMP8WI6ho='
