+++ lib/tasker.js
@@ -1,79 +1,60 @@
 /* global Zinnia */
 
-import { assertOkResponse, assertRedirectResponse } from './http-assertions.js'
+import { assertOkResponse } from './http-assertions.js'
 import { assertEquals, assertInstanceOf } from 'zinnia:assert'
 
 /** @typedef {{cid: string; minerId: string;}} RetrievalTask */
 
 export class Tasker {
-  #lastRoundUrl
   /** @type {Task[]} */
-  #remainingRoundTasks
   #fetch
 
   /**
-   * @param {object} args
-   * @param {globalThis.fetch} args.fetch
-   * @param {number} args.numTasks
-   */
+     * @param {object} args
+     * @param {globalThis.fetch} args.fetch
+     * @param {number} args.numTasks
+     */
   constructor ({
-    fetch = globalThis.fetch,
-    numTasks = 10
+    fetch = globalThis.fetch
   } = {}) {
     this.#fetch = fetch
-
-    this.numTasks = numTasks
-
-    // TODO: persist these two values across module restarts
-    // Without persistence, after the Spark module is restarted, it will start executing the same
-    // retrieval tasks we have already executed
-    this.#lastRoundUrl = 'unknown'
-    this.#remainingRoundTasks = []
   }
 
   /**
-   * @returns {Task | undefined}
-   */
-  async next () {
-    await this.#updateCurrentRound()
-    return this.#remainingRoundTasks.pop()
-  }
-
-  async #updateCurrentRound () {
+    * @param {object} args
+    * @param {number} args.roundNum
+    * @param {number} args.numTasks
+    * @returns {Promise<Task[] | undefined>}
+    */
+  async getRetrievalTasks ({ roundNum, numTasks }) {
     console.log('Checking the current SPARK round...')
     let res = await this.#fetch('https://api.filspark.com/rounds/current', {
       method: 'GET',
       headers: { 'Content-Type': 'application/json' },
-      redirect: 'manual',
       signal: AbortSignal.timeout(10_000)
     })
 
-    await assertRedirectResponse(res, 'Failed to find the URL of the current SPARK round')
-    const roundUrl = res.headers.get('location')
-    if (roundUrl === this.#lastRoundUrl) {
-      console.log('Round did not change since the last iteration')
-      return
+    let roundUrl = res.url
+    if (roundNum > 0) {
+      roundUrl = replaceRoundNumber(roundUrl, roundNum)
+      console.log('Fetching round details at location %s', roundUrl)
+      res = await this.#fetch(roundUrl, {
+        method: 'GET',
+        headers: { 'Content-Type': 'application/json' },
+        signal: AbortSignal.timeout(10_000)
+      })
+
+      await assertOkResponse(res, 'Failed to fetch the current SPARK round')
     }
 
-    console.log('Fetching round details at location %s', roundUrl)
-    res = await this.#fetch(`https://api.filspark.com${roundUrl}`, {
-      method: 'GET',
-      headers: { 'Content-Type': 'application/json' },
-      signal: AbortSignal.timeout(10_000)
-    })
-    await assertOkResponse(res, 'Failed to fetch the current SPARK round')
-    const { retrievalTasks, maxTasksPerNode, ...round } = await res.json()
+    const { retrievalTasks, ...round } = await res.json()
     console.log('Current SPARK round:', round)
-    console.log('  %s max tasks per round', maxTasksPerNode ?? '<n/a>')
     console.log('  %s retrieval tasks', retrievalTasks.length)
-    this.maxTasksPerRound = maxTasksPerNode
 
-    this.#remainingRoundTasks = await pickTasksForNode({
+    return await pickRandomTasks({
       tasks: retrievalTasks,
-      numTasks: this.numTasks,
+      numTasks
     })
-
-    this.#lastRoundUrl = roundUrl
   }
 }
 
@@ -83,7 +64,7 @@ export class Tasker {
  * @param {number} args.numTasks
  * @returns {Promise<Task[]>}
  */
-export async function pickTasksForNode ({ tasks, numTasks }) {
+export async function pickRandomTasks ({ tasks, numTasks }) {
   assertInstanceOf(tasks, Array, 'tasks must be an array')
   assertEquals(typeof numTasks, 'number', 'numTasks must be a number')
 
@@ -91,4 +72,10 @@ export async function pickTasksForNode ({ tasks, numTasks }) {
   tasks.splice(numTasks)
 
   return tasks
+}
+
+function replaceRoundNumber (roundUrl, roundNum) {
+  const parts = roundUrl.split('/')
+  parts[parts.length - 1] = roundNum
+  return parts.join('/')
 }
